moved types from 'constants' to 'types' directory
