+++ src/components/organisms/offers-list-with-map/offers-list-with-map.tsx
@@ -1,5 +1,4 @@
-import { OfferData } from '../../../types';
-import { CitiesType } from '../../../constants';
+import { OfferData, CitiesType } from '../../../types';
 import { Map, OffersListWithSort } from '../../molecules';
 import { offersActions, offersSelectors } from '../../../store/slices/offers';
 import { useActionCreators, useAppSelector } from '../../../store/helpers.ts';
