+++ src/constants/index.ts
@@ -1,12 +1,11 @@
-import CITIES from './cities.ts';
+import { CITIES } from './cities.ts';
 import { APP_ROUTE, API_ROUTE } from './routes.ts';
-import SORT_TYPES from './sort.ts';
-import AUTH_STATUS from './auth-status.ts';
+import { SORT_TYPES } from './sort.ts';
+import { AUTH_STATUS } from './auth-status.ts';
 import { URL_MARKER_CURRENT } from './map-icons.ts';
 import { URL_MARKER_DEFAULT } from './map-icons.ts';
-import { CitiesType } from './cities.ts';
-import { SortTypesType } from './sort.ts';
 import { BASE_URL } from './api.ts';
+import { REQUEST_STATUS } from './request-status.ts';
 
 export {
   CITIES,
@@ -17,9 +16,5 @@ export {
   URL_MARKER_DEFAULT,
   URL_MARKER_CURRENT,
   BASE_URL,
-};
-
-export type {
-  CitiesType,
-  SortTypesType,
+  REQUEST_STATUS,
 };
