+++ src/constants/request-status.ts
@@ -1,0 +1,11 @@
+const REQUEST_STATUS = {
+  Idle: 'Idle',
+  Loading: 'Loading',
+  Status: 'Status',
+  Error: 'Error',
+} as const;
+
+export {
+  REQUEST_STATUS,
+};
+
