+++ src/constants/sort.ts
@@ -1,5 +1,3 @@
-type SortTypesType = typeof SORT_TYPES[keyof typeof SORT_TYPES];
-
 const SORT_TYPES = {
   Popular: 'Popular',
   PriceASC: 'Price: low to high',
@@ -7,7 +5,6 @@ const SORT_TYPES = {
   Rating: 'Top rated first',
 } as const;
 
-export default SORT_TYPES;
-export type {
-  SortTypesType,
+export {
+  SORT_TYPES,
 };
