+++ src/types/cities.ts
@@ -1,0 +1,7 @@
+import { CITIES } from '../constants';
+
+type CitiesType = typeof CITIES[keyof typeof CITIES];
+
+export type {
+  CitiesType,
+};
