+++ src/utils/city.ts
@@ -1,4 +1,5 @@
-import { CITIES, CitiesType } from '../constants';
+import { CITIES } from '../constants';
+import { CitiesType } from '../types';
 
 const isCity = (value: string): value is CitiesType => (
   Object.values<string>(CITIES).includes(value)
