+++ src/utils/routing.ts
@@ -1,4 +1,5 @@
-import { APP_ROUTE, CitiesType } from '../constants';
+import { APP_ROUTE } from '../constants';
+import { CitiesType } from '../types';
 
 const createMainRouteWithCity = (city: CitiesType) => APP_ROUTE.MainWithCity.replace(':city', city.toLowerCase());
